exam 11- crud with cookie

+++ reza_exam_11/admin/list.php
@@ -1,5 +1,5 @@
 <?php
-include_once($_SERVER['DOCUMENT_ROOT'] . '/php_pondit/guest_book/config.php');
+include_once($_SERVER['DOCUMENT_ROOT'] . '/php_pondit/reza_exam_11/config.php');
 
 
 use App\GuestBook\GuestBook;
@@ -13,7 +13,7 @@ $strValidatedData = [];
 if (array_key_exists('guestbook_data', $_COOKIE)) {
     $strValidatedData = $_COOKIE['guestbook_data'];
     $storedData = unserialize($strValidatedData);
-    Debugger::debug($storedData);
+//    Debugger::debug($storedData);
 }
 
 //$guestbook = new GuestBook($validatedData);
@@ -31,35 +31,50 @@ if (array_key_exists('guestbook_data', $_COOKIE)) {
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css" rel="stylesheet"
           integrity="sha384-+0n0xVW2eSR5OomGNYDnhzAbDsOXxcvSN1TPprVMTNDbiYZCxYbOOl7+AMvyTG2x" crossorigin="anonymous">
 
-    <title>Hello, world!</title>
+    <title>cookie crud</title>
 </head>
 <body>
 <sectiion>
     <div class="container">
-        <h1 class="text-center mb-4">Guest Book</h1>
+        <h1 class="text-center mb-4 pt-5">CONTACT US</h1>
         <div class="row">
-            <div class="col-sm-5 offset-4">
+            <div class="col-sm-12">
                 <table class="table">
                     <thead>
                     <tr>
                         <th scope="col">ID</th>
-                        <th scope="col">Name</th>
+                        <th scope="col">First Name</th>
+                        <th scope="col">Last Name</th>
+                        <th scope="col">Email</th>
+                        <th scope="col">Message</th>
                         <th scope="col">Action</th>
                     </tr>
                     </thead>
                     <tbody>
                     <?php
-                    foreach ($storedData as $key => $data) {
+                    if(isset($storedData) && count($storedData) > 0){
+                        $i = 0;
+                        foreach ($storedData as $key => $data) {
+                            ?>
+                            <tr>
+                                <td><?= $i++; ?></td>
+                                <td><?= $data['firstName']; ?></td>
+                                <td><?= $data['lastName']; ?></td>
+                                <td><?= $data['email']; ?></td>
+                                <td><?= $data['message']; ?></td>
+                                <td><a href="show.php? guestsPosition=<?= $key ?>" style="text-decoration: none;">Show</a> |
+                                    <a href="edit.php? guestsPosition=<?= $key ?>" style="text-decoration: none;">Edit</a> |
+                                    <a href="delete.php?guestsPosition=<?= $key ?>" style="text-decoration: none;">Delete</a>
+                                </td>
+                            </tr>
+                            <?php
+                        }
+                    }else{
                         ?>
                         <tr>
-                            <td><?= $key; ?></td>
-                            <td><?= $data['full_name']; ?></td>
-                            <td><a href="show.php? guestsPosition=<?= $key ?>" style="text-decoration: none;">Show</a> |
-                                <a href="show.php? guestsPosition=<?= $key ?>" style="text-decoration: none;">Edit</a> |
-                                <a href="delete.php?guestsPosition=<?= $key ?>" style="text-decoration: none;">Delete</a>
-                            </td>
+                            <td colspan="3">No data is Available.</td>
                         </tr>
-                        <?php
+                    <?php
                     }
                     ?>
